Changed: Move `addListViewToDropdown()` helper to ui/dropdown/utils.
